Move all converters to converters.py
